convert predicates to int32

+++ src/eval/runtime.jl
@@ -53,7 +53,7 @@ function primitives!(mod)
   method!(mod, :tuple, RMethod(:tuple, lowerpattern(rvx"args")..., identity, true))
   method!(mod, :part, RMethod(:part, lowerpattern(rvx"(data, i)")..., part, true))
   # TODO: this is a hacky fallback
-  method!(mod, Symbol("isa?"), RMethod(Symbol("isa?"), lowerpattern(rvx"(x, T)")..., (x, T) -> tag(x) == T))
+  method!(mod, Symbol("isa?"), RMethod(Symbol("isa?"), lowerpattern(rvx"(x, T)")..., (x, T) -> Int32(tag(x) == T)))
 
   partial_widen(x::Primitive) = typeof(x)
   partial_widen(x) = x
@@ -64,7 +64,7 @@ function primitives!(mod)
     method!(mod, Symbol("isa?"), RMethod(Symbol("isa?"), lowerpattern(parse("(x, `$T`)"))..., x -> Int32(isprimitive(x, T))))
   end
   mod[:PrimitiveString] = :PrimitiveString
-  method!(mod, Symbol("isa?"), RMethod(Symbol("isa?"), lowerpattern(rvx"(x, `PrimitiveString`)")..., x -> x isa String))
+  method!(mod, Symbol("isa?"), RMethod(Symbol("isa?"), lowerpattern(rvx"(x, `PrimitiveString`)")..., x -> Int32(x isa String)))
   return mod
 end
 
@@ -79,7 +79,7 @@ function interpreter_primitives!(mod)
   method!(mod, :widen, RMethod(:widen, lowerpattern(rvx"(x,)")..., identity))
   method!(mod, :_print, RMethod(:_print, lowerpattern(parse("(a: String,)"))..., x -> (print(x); rnothing)))
   method!(mod, :string, RMethod(:string, lowerpattern(parse("(x: Symbol,)"))..., Base.string))
-  method!(mod, Symbol("isa?"), RMethod(Symbol("isa?"), lowerpattern(rvx"(x, T)")..., (x, T) -> tag(x) == T))
+  method!(mod, Symbol("isa?"), RMethod(Symbol("isa?"), lowerpattern(rvx"(x, T)")..., (x, T) -> Int32(tag(x) == T)))
   method!(mod, :(==), RMethod(:(==), lowerpattern(rvx"(x: Symbol, y: Symbol)")..., (x, y) -> Int32(x==y)))
   for T in [Int64, Int32, Float64, Float32]
     mod[Symbol(T)] = Symbol(T)
